+++ util/precomp/lut5.cpp
@@ -16,17 +16,17 @@ void do5input(uint8_t* lut5,
 
     // Decompose as 1 vs 3
     for (int i = 0; i < nLUT1; ++i) {
-        uint32_t left = lut1[i];
+        uint64_t left = lut1[i];
         for (int j = 0; j < nLut4; ++j) {
-            uint32_t right = lut4[j];
-            vector<uint32_t> llist{
+            uint64_t right = lut4[j];
+            vector<uint64_t> llist{
                     expand<5>(1, left, bin2int({1, 0, 0, 0, 0})),
                     expand<5>(1, left, bin2int({0, 1, 0, 0, 0})),
                     expand<5>(1, left, bin2int({0, 0, 1, 0, 0})),
                     expand<5>(1, left, bin2int({0, 0, 0, 1, 0})),
                     expand<5>(1, left, bin2int({0, 0, 0, 0, 1})),
             };
-            vector<uint32_t> rlist{
+            vector<uint64_t> rlist{
                     expand<5>(4, right, bin2int({0, 1, 1, 1, 1})),
                     expand<5>(4, right, bin2int({1, 0, 1, 1, 1})),
                     expand<5>(4, right, bin2int({1, 1, 0, 1, 1})),
@@ -53,10 +53,10 @@ void do5input(uint8_t* lut5,
 
     // Decompose as 2 vs 3
     for (int i = 0; i < nLUT2; ++i) {
-        uint32_t left = lut2[i];
+        uint64_t left = lut2[i];
         for (int j = 0; j < nLut3; ++j) {
-            uint32_t right = lut3[j];
-            vector<uint32_t> llist{
+            uint64_t right = lut3[j];
+            vector<uint64_t> llist{
                     expand<5>(2, left, bin2int({1, 0, 0, 0, 1})),
                     expand<5>(2, left, bin2int({1, 0, 0, 1, 0})),
                     expand<5>(2, left, bin2int({1, 0, 1, 0, 0})),
@@ -68,7 +68,7 @@ void do5input(uint8_t* lut5,
                     expand<5>(2, left, bin2int({0, 0, 1, 1, 0})),
                     expand<5>(2, left, bin2int({0, 0, 0, 1, 1})),
             };
-            vector<uint32_t> rlist{
+            vector<uint64_t> rlist{
                     expand<5>(3, right, bin2int({0, 1, 1, 1, 0})),
                     expand<5>(3, right, bin2int({0, 1, 1, 0, 1})),
                     expand<5>(3, right, bin2int({0, 1, 0, 1, 1})),
@@ -81,7 +81,7 @@ void do5input(uint8_t* lut5,
                     expand<5>(3, right, bin2int({1, 1, 1, 0, 0})),
             };
             for (int k = 0; k < 10; ++k) {
-                uint32_t l = (rlist.at(k));
+                uint64_t l = (rlist.at(k));
                 uint32_t r = (rlist.at(k));
                 uint32_t And = l & r;
                 uint32_t nAnd = ~And;
